Admin/Staff Edit Update data & Delete

+++ admin/admins.php
@@ -13,7 +13,7 @@
             <?php alertMessage(); ?>
             <?php
             $admins = getAll('admins');
-            if(!$admins){
+            if (!$admins) {
                 echo '<h4>Somethnig went wrong!</h4>';
                 return false;
             }
@@ -37,8 +37,8 @@
                                     <td><?= $adminItem['name'] ?></td>
                                     <td><?= $adminItem['email'] ?></td>
                                     <td>
-                                        <a href="admins-edit.php" class="btn btn-success btn-sm">Edit</a>
-                                        <a href="admins-delete.php" class="btn btn-danger btn-sm">Delete</a>
+                                        <a href="admins-edit.php?id=<?= $adminItem['id']; ?>" class="btn btn-success btn-sm">Edit</a>
+                                        <a onclick="return confirm('Are you sure you want to delete this record?');" href="admins-delete.php?id=<?= $adminItem['id']; ?>" class="btn btn-danger btn-sm">Delete</a>
                                     </td>
                                 </tr>
                             <?php endforeach; ?>
